requirements1.txt
# ─── Core Python / Web ───────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1

# ─── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── Environment ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Deep Learning ────────────────────────────────────────────────────────────
# TensorFlow 2.15 is the last version with good Python 3.10 + CPU-only support
tensorflow==2.15.0

# ─── Image Processing ─────────────────────────────────────────────────────────
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.24.3

# ─── AI Agent ─────────────────────────────────────────────────────────────────
groq==0.8.0

# ─── PDF Generation ───────────────────────────────────────────────────────────
reportlab==4.1.0

# ─── Data ─────────────────────────────────────────────────────────────────────
pandas==2.0.3

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.1.1
httpx==0.27.2
reportlab 
groq==0.8.0 
python-jose[cryptography]==3.3.0 bcrypt==4.0.1
passlib 
dotenv